day 5!
Today was a significant increase in difficulty. The hard part here was
all the parsing. The actual problem was pretty simple. I think I want to
revisit this one later and see if I can clean up this solution to use an
array rather than the stack implementation. I would like to try to get
better with String.splice().