Day 6!
I came up with a generic solution to this one :)
I really should have been faster on this one since it was actually
rediculously easy. I was trying to be fancy using slice and a fancy way
of checking for uniqueness instead of the simple solution I ended up
with.