Day 11!
Part One was a ton of fun. I had to use math for part two which almost
made me go to bed as soon as I realized. The sample size of the monkeys
was small enough that I could calculate a modulo to use by taking the
product of all the different values used to test. I cleaned up my
solution after because it was pretty messy from when I converted it to
bigint and then realized I would need math to make this computationally
possible. This math works because the mod I am using is divisible by
every number that will be used to test.