Day 13!!
oof, this one was super annoying. Spent forever trying to parse it
myself and eventually decided to use the JSON parser cuz I saw memes on
redit. I am really curious how typescript sorting works under the hood
because I implemented a cache for my compare function and it was never
called so I ended up trashing it. I did verify that the cache was
working if the same call was made twice. I wonder if the sort method
does it's own caching.