Day 16!!
What a tough one! My first solution to part 2 took around an hour to
run. While that was happening, I reworked part one to use numbers for
the nodes which would allow me to represent the path as a single number
and treat each node as a bit in that path. This made part two way faster
for two reasons.
1. The paths no longer cared about the order, just the best path with a
   given set of nodes. This reduces the size of the array by a factor of
   10.
2. The comparisons between two paths is much faster. I think this is a
   benefit, but no where near the benefit realized by reducing the size
   of the array.