Day 17!!
I am a day late and really need sleep. This one was quite a challenge
and I was definitely making it harder on myself by only looking for
patterns at the end of the wind input which was not successful. After I
rewrote it to use a map, it worked.