Day 18!!
Today was pleasant compared to the last few days. I am not sure if there
is a better way to do it than brute forcing all the open edges from the
outside in, but my solution ran pretty fast so I am happy with it. All
the troubles I ran in to today were pretty simple to solve and it made
it much more enjoyable.