Day 19 :(
I definitely need to come back and take a look at this. I got the
answers, but I know there is a bug in my code because I had to change
the pruning parameters (make them worse) when I went to part two in
order to get the right answer so I know there is a bug somewhere.