Day 21!!
Today was actually a lot of fun again :)
Part one was a breeze and part two wasn't too harc either. I did not try
and actually solve the equation. Just used the guess and check method on
steroids. Luckily there was a pattern I could use. I actually had an
intermediate version of part two where it prompted the user for the
number and I tested my binary search algorighm on the example.