Day 23!!
Not too bad, especially when compared to the debugging required for
yesterday. I did have an annoying bug that was preventing my part two
from working and took me forever to find. I was calling something as the
second part of an or clause so it was only getting ran once.